requirements.txt
pyperclip==1.8.2
pystray>=0.19.4
Pillow>=9.0.0
pyinstaller>=5.0.0
pynput>=1.7.6
